+++ benchmarks/skewed_workload.py
@@ -36,7 +36,8 @@ for read_percent in range(11):
     for run in range(TRIALS):
         print "Skewed Workload (Read percent: %s, Trial no.: %s)" % (read_percent, run)
 
-        shutil.rmtree(DB_LOCATION)
+        if os.path.exists(DB_LOCATION):
+            shutil.rmtree(DB_LOCATION)
         shutil.copytree(DB_EXISTING, DB_LOCATION)
 
         # Run control
@@ -52,7 +53,8 @@ for read_percent in range(11):
         f.close()
         os.remove('example.txt')
 
-        shutil.rmtree(DB_LOCATION)
+        if os.path.exists(DB_LOCATION):
+            shutil.rmtree(DB_LOCATION)
         shutil.copytree(DB_EXISTING, DB_LOCATION)
 
         # Run experimental
