+++ benchmarks/static_workload.py
@@ -35,7 +35,8 @@ for i in range(11):
     for run in range(TRIALS):
         print "Static Workload (Read percent: %s, Trial no.: %s)" % (read_percent, run)
 
-        shutil.rmtree(DB_LOCATION)
+        if os.path.exists(DB_LOCATION):
+            shutil.rmtree(DB_LOCATION)
         shutil.copytree(DB_EXISTING, DB_LOCATION)
 
         # Run control
@@ -51,7 +52,8 @@ for i in range(11):
         f.close()
         os.remove('example.txt')
 
-        shutil.rmtree(DB_LOCATION)
+        if os.path.exists(DB_LOCATION):
+            shutil.rmtree(DB_LOCATION)
         shutil.copytree(DB_EXISTING, DB_LOCATION)
 
         # Run experimental
